+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: nfelsemb <nfelsemb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/24 13:08:16 by nfelsemb          #+#    #+#             */
-/*   Updated: 2021/11/26 12:31:06 by nfelsemb         ###   ########.fr       */
+/*   Updated: 2021/11/26 14:10:12 by nfelsemb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int		ft_isascii(int c);
 int		ft_isdigit(int c);
 int		ft_isprint(int c);
 void	*ft_memset(void *s, int c, size_t n);
-int		ft_strlen(char *tab);
+size_t	ft_strlen(char *tab);
 int		ft_tolower(int c);
 int		ft_toupper(int c);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
